:recycle: [Refactor] query문 fundingRepo -> contributorRepo로 이동
