+++ src/main/java/kcs/funding/fundingboost/domain/repository/Contributor/ContributorRepository.java
@@ -1,4 +1,4 @@
-package kcs.funding.fundingboost.domain.repository;
+package kcs.funding.fundingboost.domain.repository.Contributor;
 
 import java.util.List;
 import kcs.funding.fundingboost.domain.entity.Contributor;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ContributorRepository extends JpaRepository<Contributor, Long> {
+public interface ContributorRepository extends JpaRepository<Contributor, Long>, ContributorRepositoryCustom {
 
     @Query("select c from Contributor c" +
             " join fetch c.funding f" +
