+++ src/main/java/kcs/funding/fundingboost/domain/repository/Contributor/ContributorRepositoryCustom.java
@@ -1,0 +1,6 @@
+package kcs.funding.fundingboost.domain.repository.Contributor;
+
+public interface ContributorRepositoryCustom {
+
+    Long countContributorsForFunding(Long fundingId);
+}
