+++ src/main/java/kcs/funding/fundingboost/domain/repository/funding/FundingRepositoryCustom.java
@@ -5,13 +5,9 @@ import kcs.funding.fundingboost.domain.entity.Funding;
 
 public interface FundingRepositoryCustom {
 
-    List<Funding> findAllByMemberIn(List<Long> memberIds);
-
     Funding findFundingInfo(Long memberId);
 
     Funding findMemberByFundingId(Long fundingId);
 
     List<Funding> findFundingByMemberId(Long memberId);
-
-    Long countContributorsForFunding(Long fundingId);
 }
